requirements.txt
web3==7.7.0
Django==5.1.5
djangorestframework==3.15.2
gunicorn==23.0.0